Define getLocation

+++ src/main/kotlin/org/example/operation_quasar_fire/model/entities/Position.kt
@@ -1,3 +1,3 @@
 package org.example.operation_quasar_fire.model.entities
 
-data class Position(val x: Int, val y: Int)
+data class Position(val x: Float, val y: Float)
